+++ src/cmds/mod.rs
@@ -1,2 +1,3 @@
 pub mod ai;
+pub mod config;
 pub mod new;